Upgrade from Junit 4 to Junit 5
This commit achieves two things 1) it moves Junit dependency declarations out of subprojects and into the root project, to reduce redundant declarations that are difficult to maintain 2) it upgrades the version of Junit used from version 4 to 5. However, this upgrade still relies on backwards compatibility features of Junit 5, to continue to run existing tests.