Change secp256k1 provider from SunEC to Bouncy Castle
Currently, the code relies on the native SunEC crypto provider for the implementation of the secp256k1 curve. As of JDK15, this implementation has been disabled. This commit migrates to using Bouncy Castle as the replacement provider of this implementation. The commit also upgrades to the latest version of Bouncy Castle